+++ LiteLoader.NET/Header/Logger/Logger.hpp
