+++ LiteLoader.NET/Main/GlobalClass.cpp
@@ -1,5 +1,6 @@
 #include "DotNETGlobal.hpp"
 
+#include <LiteLoader.NET/Header/Logger/Logger.hpp>
 
 inline void GlobalClass::Init()
 {
