+++ LiteLoader.NET/Main/Info.hpp
@@ -17,9 +17,9 @@
 
 #define LLNET_VERSION_MAGOR 2
 
-#define LLNET_VERSION_MINOR 4
+#define LLNET_VERSION_MINOR 5
 
-#define LLNET_VERSION_REVISION 3
+#define LLNET_VERSION_REVISION 0
 
 #define LLNET_VERSION_BUILD 0
